REGRESSION(260575@main): TestWebKitAPI.AudioRoutingArbitration.Deletion times out with UI-side-compositing enabled https://bugs.webkit.org/show_bug.cgi?id=253670 rdar://106322713
Reviewed by Ryosuke Niwa.

When we did the "don't host layers from the WebContent process" work, we re-used the
PlaybackSessionManager/VideoFullscreenManager codepath from element fullscreen and
picture-in-picture to handle creation and sizing of the video layer in the UI process. This caused
a pre-existing bug where a fullscreened HTMLMediaElement would never be destroyed (due to having
eventListeners) to now apply to all HTMLMediaElements displayed in a page.

The eventListeners aren't removed because a boolean ivar intended to track whether eventListeners
were added was never set to true. Once that ivar is correctly set, the test passes.

* Source/WebCore/platform/cocoa/PlaybackSessionModelMediaElement.mm:
(WebCore::PlaybackSessionModelMediaElement::setMediaElement):

Canonical link: https://commits.webkit.org/261503@main