[macOS] Assertion failure in nonVisibleProcessMemoryCleanupTimerFired https://bugs.webkit.org/show_bug.cgi?id=253743 rdar://problem/106579817
Reviewed by Chris Dumez.

Using nonVisibleProcessMemoryCleanupTimer is not a good idea for
clearing memory after suspension. If the page is still in the window
then nonVisibleProcessMemoryCleanupTimerFired will hit an assertion. It
also does not account for resuming processes before the timer fires.
Disable the behavior for now to fix the assert failure while a better
solution is developed.

* Source/WebKit/WebProcess/WebProcess.cpp:
(WebKit::WebProcess::prepareToSuspend):

Canonical link: https://commits.webkit.org/261537@main