[GTK][WPE] Gardening `fast/inline/text-decoration-underline-inline-box-offset.html`
Unreviewed test gardening.

The result of the previous version of this test depended on the
descending of the font to which `Monospace` resolved on the platform.

To fix this, we now use a web font, which all platforms should pick up.

The test also didn't really test vertical text because the style for it
was identical in both the test and the baseline files.

The idea is to set the bottom border for horizontal text and the left
border for vertical text so that it matches the geometry of the
underline.

*
LayoutTests/fast/inline/text-decoration-underline-inline-box-offset-expected.html:
*
LayoutTests/fast/inline/text-decoration-underline-inline-box-offset.html:

Canonical link: https://commits.webkit.org/261565@main