Concat with a CSSStyleSheet and shadowRoot.adoptedStyleSheets returns array in array https://bugs.webkit.org/show_bug.cgi?id=253814
Reviewed by NOBODY (OOPS!).

Per the EcmaScript specification for Array.prototype.concat [1], it should
"spread" the second parameter if isArray() returns true. However, our
implementation would only do so if isJSArray() returns true for the second
parameter. This was causing concat() to not do the right thing when the
second parameter was a JSObservableArray [2].

In this patch, I kept the JSArray code path as a fast path and add support for
array-like objects as a slow path.

[1] https://tc39.es/ecma262/multipage/indexed-collections.html#sec-array.prototype.concat
[2] https://webidl.spec.whatwg.org/#es-observable-array

* LayoutTests/cssom/array-concat-adoptedStyleSheets-expected.txt: Added.
* LayoutTests/cssom/array-concat-adoptedStyleSheets.html: Added.
* Source/JavaScriptCore/runtime/ArrayPrototype.cpp:
(JSC::concatAppendOne):
(JSC::concatAppendArrayLike):
(JSC::JSC_DEFINE_HOST_FUNCTION):