gcGLSpanDynamicExtent is defined in incorrect header, GCGLSpan.h included in wrong order https://bugs.webkit.org/show_bug.cgi?id=253887 rdar://problem/106701929
Reviewed by NOBODY (OOPS!).

Include GCGLSpan.h in the position where the project coding guidelines
tell.

gcGLSpanDynamicExtent is part of GCGLSpan.h, and that's where it should
be defined.

Default values for template parameters of forward declared types should
be defined as they need to be. gcGLSpanDynamicExtent cannot be defined
in one forward declaration site; this is not logical. If there was two
forward declare sites, then what?

* Source/WebCore/html/canvas/WebGLRenderingContextBase.h:
* Source/WebCore/platform/graphics/GCGLSpan.h:
* Source/WebCore/platform/graphics/GraphicsContextGL.h: