REGRESSION(261586@main): The assertion is randomly failing https://bugs.webkit.org/show_bug.cgi?id=253875 rdar://106705387
Unreviewed, drop assertion in ~Document() to make sure that m_ranges is empty
as it is hitting on the bots since we updated m_ranges to be a WeakHashSet.
Ranges no longer remove themselves in their destructor from this set so this
assertion is no longer guaranteed to hold. The WeakHashSet will empty itself
as Range objects get destroyed but ~Document() may get called while the
destructor of a Range is still running.

* Source/WebCore/dom/Document.cpp:
(WebCore::Document::~Document):

Canonical link: https://commits.webkit.org/261634@main