Scripts: Use webbrowser package in Python to open a URL in a web browser https://bugs.webkit.org/show_bug.cgi?id=253515
Reviewed by Jonathan Bedard.

I made a mistake. Because "start" worked on the command line, I assumed
that it would work under the Terminal.run() command. Now I realize this
is not what actually happens. Instead of reverting that commit, commit
https://commits.webkit.org/261332@main, I am instead introducing a
proper fix, which is using the URL and opening it in a browser via the
webbrowser package.

* Tools\Scripts\libraries\webkitcorepy\webkitcorepy\terminal.py: Change
  'start' back to 'explorer.' Teach open_url to call the webbrowser
  url method to open webpages if the scheme is http or https.

* Tools\Scripts\webkitpy\common\system\user.py: Inline redundant method.

Canonical link: https://commits.webkit.org/261687@main