It's too easy to open multiple ARQL previews by mistake https://bugs.webkit.org/show_bug.cgi?id=253995 rdar://106760359
Reviewed by NOBODY (OOPS!).

The new approach for opening ARQL means that if a user taps on a link multiple
times many previews will open. The idea here is to give ARQL a way to identify
the incoming URL without telling them exactly what it is (in this case, pass a
hash of the URL to ARQL). That way they can choose to not open a new instance.

* Source/WebKit/UIProcess/Cocoa/LegacyDownloadClient.mm:
(WebKit::LegacyDownloadClient::legacyDidStart): Send the originating URL to
SystemPreviewController.
* Source/WebKit/UIProcess/Cocoa/SystemPreviewControllerCocoa.mm: Remove some
of the older SPI since they have long since been removed. Create a hash
of the URL and send it to ARQL.
* Source/WebKit/UIProcess/SystemPreviewController.h: Add a new parameter for
the download URL.